New version with lazy loading and dynamic files
Remove isotope since it was creating:
issues on lazy loading
issues on mobile
issues with animations
Added:
JS functions for filter the portofolio
Modify HTML and JS to generate dynamically all the links and images.
Create a python script which generates a JSON file to mimic a server like list of content for the grid.